requirements.txt
yfinance>=0.2.31
openai>=1.12.0
flask>=3.0.0
python-dotenv>=1.0.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
schedule>=1.2.1
plotly>=5.18.0
pandas>=2.1.5
jinja2>=3.1.3
lxml>=5.1.0
markdown>=3.5.2
apscheduler>=3.10.4
